embedding-service/requirements.txt
sentence-transformers==2.3.1
grpcio==1.60.0
grpcio-tools==1.60.0
torch==2.1.2
numpy==1.24.3
